yolo_example/requirements.txt
torch
ultralytics
stadle-client